Feat: 스케치북 레이아웃 컴포넌트 추가 및 적용
Refs: #249
