+++ src/pages/editReviewPage/EditReviewStyle.tsx
@@ -1,59 +1,9 @@
 import { css, styled } from 'styled-components';
 
 import emptyImage from '../../assets/add-image-defalut.svg';
-import ring from '../../assets/icons/ring.svg';
 import tape from '../../assets/icons/tape.svg';
 import pin from '../../assets/icons/thumbtack.svg';
 import px2vw from '../../utils/px2vw';
-
-export const PageWrapper = styled.main`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 60px ${px2vw(142)} 0 ${px2vw(144)};
-`;
-
-export const RingsWrapper = styled.section`
-  display: flex;
-  gap: 70px;
-  margin-bottom: -43px;
-  position: relative;
-  z-index: 9;
-`;
-
-export const Rings = styled.div`
-  width: 79px;
-  height: 81px;
-  background-image: url(${ring});
-  background-size: 79px 81px;
-`;
-
-export const ContentBox = styled.form`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background-color: #fffbe2;
-  padding: 81px ${px2vw(49)} 24px ${px2vw(45)};
-  border: 2px solid #1e232c;
-  border-radius: 20px;
-  position: relative;
-  &::after {
-    content: '';
-    display: block;
-    box-sizing: border-box;
-    width: 100%;
-    height: 100%;
-    background-color: #fffbe2;
-    border: 2px solid #1e232c;
-    border-radius: 20px;
-    position: absolute;
-    top: 10px;
-    left: 14px;
-    z-index: -9;
-  }
-`;
 
 export const TopSection = styled.section`
   width: 100%;
